backend/src/modules/users/services/ResetPasswordService.ts
import { injectable, inject } from 'tsyringe';

import { differenceInHours } from 'date-fns';

import IUsersRepository from '../repositories/IUsersRepository';
import IUserTokensRepository from '../repositories/IUserTokensRepository';
import IHashProvider from '../providers/HashProviser/models/IHashProvider';

import AppError from '../../../shared/errors/AppError';

interface IRequest {
  token: string;
  password: string;
}

@injectable()
class ResetPasswordService {
  constructor(
    @inject('UsersRepository')
    private usersRepository: IUsersRepository,

    @inject('UserTokensRepository')
    private userTokensRepository: IUserTokensRepository,

    @inject('HashProvider')
    private hashProvider: IHashProvider,
  ) {}

  public async execute({ token, password }: IRequest): Promise<void> {
    const userToken = await this.userTokensRepository.findByToken(token);

    if (!userToken) {
      throw new AppError('User Token does not exists', 404);
    }

    const user = await this.usersRepository.findById(userToken.user_id);

    if (!user) {
      throw new AppError('User does not exists');
    }

    const tokenCreateAt = userToken.created_at;

    if (differenceInHours(Date.now(), tokenCreateAt) > 2) {
      throw new AppError('Token expired');
    }

    user.password = await this.hashProvider.generateHah(password);

    await this.usersRepository.save(user);
  }
}

export default ResetPasswordService;
